fix: Remove auto-discovery of sessions from multiple scanners
Since we have a dedicated scanner for identifying available sessions it
does not make sense to implement a low-effort variant of
sessions-scanning for multiple scanners with totally different tasks.

Moreover, this change allows to run these scanners without requiring the
availability of DiagnosticSessionControl, as long as no --sessions flag
is given.