fix(dumpcap): Handle dumpcap invocation for unix transports
Currently there is a flaw in the plugin interface. Plugins cannot
specify how dumpcap should be invoked to capture traffic. The default
is ethernet traffic. There are Unix transports available which are
neither CAN traffic nor ethernet traffic. In this case the wrong
method `_eth_cmd()` is called which cannot create a valid CLI for
dumpcap.