feat(doip-discover): Dynamically adapt RA Request scans
It is up to the DoIP entity to decide whether it first checks for
correct SourceAddress or RoutingActivationType (while the standard
actually puts the check for SourceAddress first...).
This commit minimizes the search space by determining which check is
performed first, and accordingly adapting the following scans.
By doing so it also removes code duplication.