feat(win32): Add tier-2 support for windows platforms
There are some situations out there where contact to windows platforms
cannot be avoided. An example for this are very special setups using
vector tooling and hardware. Unfortunately, these companies support
windows only (and WSL2 is not an option, too). Windows at least provides
something like a commandline and gallia is able to run natively; but
with some restrictions:

- pcap creation is not supported
- LATEST symlinks are not supported
- all protocols that use the linux can stack are not supported
- Some cli invocation might not work as expected due to special handling
  withing the windows terminal. For instance, use the `gallia.toml` file
  to avoid frustration with `--target`.

This patch adds, let's call it tier-2 support for windows. My
definition of tier-2 is: We guarantee that gallia does not crash due to
the usage of unavailable platform APIs. In other words, if a feature does
not work on Windows it will be disabled rather than ported.

There is a linter job available for checking those cases.