feat(flexray): Add support for FlexRay via the Vector XL library
This patch adds FlexRay support using the Vector XL library. This only
works on windows. Since there is no general software stack available on
windows, an implementation for ISO-TP over FlexRay is included in this
patch.

In theory, the ISO-TP implementation can be used for CAN as well. If
needed it can be placed in its own module after some refactoring.
However, it is not complete: FlowControl is not implemented completely.
In my tests it did work for the test setup I had.