fix: Remove pygit2 dependency
On a few platforms (ask @peckto) the pygit2 dependency cannot be
installed with poetry, since the toolchain tries to build from source
and fails. Since we only use this library to detect the git root, let's
implement this with calling the Git binary instead. When git is not available,
then Git repos are not used for config discovery.