chore: Remove exitcodes dependency and vendor it
[Russ Cox](https://research.swtch.com/deps) often blogs about the need
of dependencies. He argues, that if the maintenance cost is low, then it
is better to avoid the dependency at all and just copy the the code.

Since the ultimate goal for gallia should be the inclusion in larger
Linux distributions the fewer dependencies we have the better. This one
for sure is avoidable and easily vendored.