fix: Remove dependence on special Python builds for CAN functionality
The CAN constants are compiled in depending in the enabled kernel
features. In case a Python interpreter is built in a special
environment (e.g., https://github.com/indygreg/python-build-standalone
used by the uv tool), then a few constants might be missing. The absence
of these constants does not limit the availability of the CAN feature on
a supported kernel. So, lets checkin a generated file with the correct
constants to avoid running into this problem.