chore: Remove aiofiles dependency
We only write very small files. Unless the artifactsdir is located on a,
e.g., very slow network drive, it is very unlikely that the eventloop is
blocked by these writes. Usually only a few bytes are written to
dumpfiles or to separate result files.

Let's use the stdlib instead and get rid of a further dependency.