feat(scan-sessions): Greatly speed up scan by recovering stack on need
The current approach does recover the stack on each session change
attempt, which greatly prolongs scans in a session that takes some time
to get into.
From now on, the stack is only recovered when needed, e.g. in case of an
error or when a session change attempt was successful.