!feat(scan-sessions): Invert scanning approach: --fast to --thorough
With the assumption that it does not matter how a session is entered, it
is only interesting to know from where one can reach a session.
Thus, the better default is to use a fast approach that only scans from
each session once in order to find available session transitions.

To replicate the old scanning behaviour, --thorough can be specified to
scan multiple times from the same session.
In order to avoid infinite loops, --depth now has a default value of 4.