fix(client): Do no longer send TesterPresent in ResponsePending loop
We now had multiple occasions where the TesterPresent request with
suppress response bit set triggered a reply by the ECU.
This reply will be treated as the awaited answer of the ResponsePending
loop which cannot be recovered from.

Not sending TesterPresent in the loop also conforms to the UDS standard.