fix(db): Create single task that executes DiagnosticMessage queries
This commit introduces an executor task that handles queries coming from
a queue. Compared to the old approach that easily generated >100k tasks
per scan, this approach is way more memory friendly.

What adds to this is that previously the tasks were generated in a
different event loop than the one they were awaited in, leading to
dangling objects that persisted until the python interpreter closed.
This commit removes these memory leaks which becomes extremely important
when scanners are used "as lib".