feat: Add a dowloader for the goepel g-api
This is the proprietary software world… We are not allowed to ship this
thing (to create the best user experience), but we are allowed to
download it…

Best: Only a creepy self extracting bash script is available that
requires root permissions and so on. Just run this in a container and
grab the extracted files.