!feat: Remove HAS_ARTIFACTS_DIR in favor of artifacts_base config
Currently, the property `HAS_ARTIFACTS_DIR` of `BaseCommand` determines
whether a folder is created and artifacts are saved.

This commit removes this property in favor of the config property
`artifacts_base`, such that artifacts are automatically saved, once the
property is filled in the config object.

These adaptions bring two major benefits:

1. If gallia is run quickly from the CLI, it does not clutter temporary
   storage or require dependencies such as dumpcap.
2. When calling individual scanners from within python with gallia as a
   lib, the interface to decide whether and where to artifacts are saved
   is more straightforward.