fix(uds): Properly raise RequestResponseMismatch errors
`RequestResponseMismatch` errors will be given priority over
`MalformedResponse` errors, because they violate the standard even
harder.

Additionally, this commit fixes a bug where `RawRequest` could never
pass the comparison to a positive reply, which results in the odd
behavior that scanning services might result in errors when a service
actually responds with a positive reply instead of a negative one.