fix(uds): Shield TesterPresent from being cancelled mid-request
This commit prevents TesterPresent worker from being cancelled midway
through the .ping method, which might result in half-handled requests
that lead to RequestResponse mismatches.

This is not unlikely, but actually happens each time any response of the
ECU is sent more than 500ms after the initial request and the worker is
cancelled immediately after, as TesterPresent's ping method then is
already waiting for the mutex to be released to send 3e00.