fix(uds): Shield all UDS-requests from being cancelled mid-request
This commit prevents UDS-requests from being cancelled midway through
the `_request` method, which might result in half-handled requests that
lead to RequestResponse mismatches.

This change becomes relevant if multiple scanner objects exists in
parallel and an unclean shutdown (CancelledError) is introduced.

E.g., for TesterPresent, this is not unlikely, but actually happens each
time any response of the ECU is sent more than 500ms after the initial
request and the worker is cancelled immediately after, as
TesterPresent's ping method then is already waiting for the mutex to be
released to send 3e00.