chore(dumpcap): Attach dumpcap to BaseTransport
This commit links the closely connected capture mechanism with Dumpcap
directly to the `BaseTransport` class, where it thematically belongs.

This allows anybody who uses `Basetransport` or its childs to easily
start/stop capturing. An example for this is the DoIP discoverer, which
currently has no dumpcap features because of it being an `AsyncScript`.