feat(isotp-discover): Automatically scan with and without padding
Some ECUs might require padding, while others don't. It does not cost
much to send each message both with and without padding automatically,
to cover both cases.
If a specific padding is given on the cli, only this padding is used.

Only the first occurrence of a ID is saved, i.e. without padding if it
works without.

A similar approach was also suggested by @mich41v4294 in #527.